[c++] Move two headers and adjust two include statements
